Add additional SSH config options to scheduler (#34)
This PR adds a few additional config fields for specifying ssh keys and the ssh user for creating and provisioning nodes. As a part of this PR I had to expose an additional field in the driver-specific configuration called extra_creation_kwargs which can get passed to libcloud's create_node() method.