Remove locking in watch_job_task
There is currently a race condition between watch_job_task and
scale_job_task which can result in the job failing due to it
scaling at the same time that the scheduler starts watching it.
This needs to be investigated further, but in the meantime this
commit disables the locking step when watching since this step
was primarily a performance optimization (to limit scenarios where
redundant tasks are running for the same job).