Update job state lifecycle management in watch and scale job tasks
Updated scale_job_task to require it to aquire a lock on the job
record in the database. This prevents scenarios where the controller
requests scaling before start_job_task has committed, leading to
extra nodes being created (which can cause MPI to crash with cryptic
SSH errors).

Also updated watch_job_task to not be concerned with updating the
watched job's state in the database since the downstream
stop_job_task already handles this and made it so that the task is
retried only in cases of GRPC errors.