Infrastructure overhaul: Terraform + Helm charts (#363)
This also now includes all containers that have previously been managed separately using docker-compose (scheduler, celery worker, ...) in Kubernetes.

* Initialize helm chart

* [WIP] Update templates for postgres, redis, and graphite

* Add terraform for main k8s secrets

* Remove lorri envrc

* Add graphite configmap

* Update terraform secrets

* [WIP] Add initial service account for scheduler to use

* [WIP] helm components for scheduler and workers

* Add local node_modules to gitignore

* Make additional fields configurable in helm chart

* Update scheduler authorization logic

Updated scheduler authorization to:

  1. Only load the Firebase client when not in dev mode
  2. Assume a constant user_id when in dev mode which can be specified
     via an environment variable.

* Allow incluster config for k8s nodes

* [WIP] Global permissions for scheduler k8s service account

* Update k8s secrets

* Update mpi4py version to allow support for newer versions of setuptools

* Add extra schedule config fields for storage backend

* Update config secrets

* [WIP] Use boto3 for zipping and fetching results

* Update config secrets

* WIP

* [WIP] Hacking...

* [WIP] Functioning job

Zip task still fails

* Fix issue with results.zip naming

* [ci skip] Terraform for GCP envs

* Update README

* [ci skip] Expose graphite and sheduler on internal VPC

* Update helm chart

* Make ImagePullPolicy configurable for job nodes

* Fix storage bucket name for dev cluster config

* Add .dockerignore

* [ci skip] Update Makefile

* [ci skip] Cluster tweaks and internal DNS

* Made some small adjustments to autoscaling config to enable
  autoprovisioning of nodes.
* Added support for an Ingress to the backend so that we can
  access backend resources via a FQDN in App Engine.

* Move regions

* [ci skip] Additional region update

* Tweaks

* Add mcmc-stats to helm chart

* Add mcmc-stats to ingress controller in dev cluster

* Update README

* Add helpers for adding and removing user permissions to scheduler

* Update deployment docs

* Update helm values files

* Add push-images targe to Makefie

* Update README

* Remove old k8s config

* Clean up docker directory

* Remove check for job state in job.scale()

* Ensure services are deleted and remove debugging lines

* Generate app.yaml with Terraform

* Update docs

* Remove outdated unit test

* Expand terraform version bound

* Add new dependencies to shell.nix

* Fix hardcoded path in zip_results_task

* Update Docker image prefixes

* Add instructions for restarting deployments

* Add Terraform lockfile for local cluster

* Fix error in entrypoint from manual conflict resolution

* Clean up dev debugging logging statements

* Revert merge of recent changes to k8s_pod.py

Reverts the state of k8s_pod.py to that of this branch. Merging
the latest state on master (which separates logic for creating
controller and worker nodes) results in cryptic kubernetes secret
failures.

* Remove locking in watch_job_task

There is currently a race condition between watch_job_task and
scale_job_task which can result in the job failing due to it
scaling at the same time that the scheduler starts watching it.
This needs to be investigated further, but in the meantime this
commit disables the locking step when watching since this step
was primarily a performance optimization (to limit scenarios where
redundant tasks are running for the same job).

* Fix failing test in chainsail.common

* Upgrade black in CI runner

* Fix failing test in chainsail.scheduler

* Run formatter

* Add missing arguments to test in test_k8s

* Place version bound on MarkDown

Newer versions have a breaking API change which is incompatible with
the version of flask we currently use.

* Fix failing tests in chainsail.scheduler

* Move from kubernetes_ingress to kubernetes_ingress_v1

* Complete GCP deployment documentation

* Update job state lifecycle management in watch and scale job tasks

Updated scale_job_task to require it to aquire a lock on the job
record in the database. This prevents scenarios where the controller
requests scaling before start_job_task has committed, leading to
extra nodes being created (which can cause MPI to crash with cryptic
SSH errors).

Also updated watch_job_task to not be concerned with updating the
watched job's state in the database since the downstream
stop_job_task already handles this and made it so that the task is
retried only in cases of GRPC errors.

* Increase number of retries for controller health check task

Co-authored-by: Simeon Carstens <simeon.carstens@tweag.io>