Make (zip) file upload work by parsing multipart/form-data
We need to add authentication headers to the request in the API
middleware. This worked so far for the application/json content type
and (uselessly) multipart/form-data with a text- / non-binary file.
To make this work for zip files, this parses the multipart/form-data
request with an external library and reconstructs a FormData object
that is then sent along with the authentication headers to the
backend.
There is probably a better way to do this for someone who actually
knows JavaScript and stuff.