Update nixpkgs to follow nixpkgs/release-22.11
The previous use of "21.11" seemed to keep the pin rather old. The
"21.11" is a tag and "release-21.11" is a branch. My guess is that
"21.11" is a tag from when the 2.11 release is branched from.

To be clear, here I've updated to to follow 22.11, not 21.11.

This gives us poetry2nix 1.39.1 and poetry 1.3.0.