Bump `pyyaml` version and unpin `mpi4py` dependency (#486)
Our previous `pyyaml` version constraint leads to a `pyyaml` version
that is incompatible with Cython
3.0.0 (https://github.com/yaml/pyyaml/issues/601). Likewise, we had
`mpi4py` pinned to a specific commit that fixes an issue with
`setuptools`, but now that commit is incompatible with Cython
3.0.0. The latest `mpi4py` release (3.1.4) works just fine, though, so
we demand at a minimum that version. `pyyaml` seems to work fine for
versions >= 6.0.0, so we set that version constraint wherever `pyyaml`
is required.